Split SerialiseTextEnvelope module out of Cardano.Api.Typed
